@@ -1,47 +1,41 @@
 package net.vexelon.currencybg.srv;
 
-import java.io.File;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-
+import com.google.common.base.Charsets;
 import org.apache.commons.configuration2.PropertiesConfiguration;
 import org.apache.commons.configuration2.builder.ReloadingFileBasedConfigurationBuilder;
-import org.apache.commons.configuration2.builder.fluent.FileBasedBuilderParameters;
 import org.apache.commons.configuration2.builder.fluent.Parameters;
 import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.apache.commons.configuration2.reloading.PeriodicReloadingTrigger;
 
-import com.google.common.base.Charsets;
+import java.io.File;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Global server configurations
- *
  */
 public enum GlobalConfig {
 	INSTANCE;
 
 	private ReloadingFileBasedConfigurationBuilder<PropertiesConfiguration> builder;
-	private PeriodicReloadingTrigger trigger;
+	private PeriodicReloadingTrigger                                        trigger;
 
 	/**
 	 * Available configuration options to read/write
-	 *
 	 */
 	public enum Options {
+		SERVER_NAME("server.name"),
 		TIMEZONE_SERVER("timezone"),
-		MAINTENANCE_ENABLED("maintenance.enabled"),
 		TELEGRAM_BOT_TOKEN("telegram.bot"),
 		TELEGRAM_CHANNEL("telegram.channel"),
-		ENABLE_LOG_SQL("log.sql"),
 		ENABLE_LOG_DEBUG("log.debug"),
 		SPARKPOST_API_KEY("sparkpost.apikey"),
 		SPARKPOST_EMAILS("sparkpost.emails"),
 		SPARKPOST_SUBJECT("sparkpost.subject"),
 		SPARKPOST_FROM("sparkpost.from"),
-		SERVER_NAME("server.name"),
 		REPORT_TYPE("reporter.type");
 
-		private String optName;
+		private final String optName;
 
 		Options(String name) {
 			this.optName = name;
@@ -60,17 +54,15 @@ public enum GlobalConfig {
 			builder = createConfigurationBuilder(file, executor);
 
 			// defaults
+			setServerName("");
 			setServerTimeZone(Defs.DATETIME_DEFAULT_TIMEZONE);
-			setMaintenanceEnabled(false);
 			setBotToken("");
 			setBotChannel("");
-			setLogSqlEnabled(false);
 			setLogDebugEnabled(false);
 			setSparkPostAPIKey("");
 			setSparkPostEmails("");
 			setSparkPostSubject("");
 			setSparkPostFrom("");
-			setServerName("");
 			setReportType("");
 
 			builder.save();
@@ -80,11 +72,6 @@ public enum GlobalConfig {
 		}
 	}
 
-	/**
-	 * 
-	 * @param file
-	 * @param executor
-	 */
 	public void load(File file, ScheduledExecutorService executor) {
 		builder = createConfigurationBuilder(file, executor);
 		builder.setAutoSave(true);
@@ -97,17 +84,13 @@ public enum GlobalConfig {
 	}
 
 	private ReloadingFileBasedConfigurationBuilder<PropertiesConfiguration> createConfigurationBuilder(File file,
-	        ScheduledExecutorService executor) {
-
-		FileBasedBuilderParameters parameters = new Parameters().fileBased().setFile(file)
-		        .setEncoding(Charsets.UTF_8.name());
-
-		ReloadingFileBasedConfigurationBuilder<PropertiesConfiguration> builder = new ReloadingFileBasedConfigurationBuilder<PropertiesConfiguration>(
-		        PropertiesConfiguration.class).configure(parameters);
+			ScheduledExecutorService executor) {
+		var parameters = new Parameters().fileBased().setFile(file).setEncoding(Charsets.UTF_8.name());
+		var builder = new ReloadingFileBasedConfigurationBuilder<>(PropertiesConfiguration.class).configure(parameters);
 
 		// setup reloading interval
 		trigger = new PeriodicReloadingTrigger(builder.getReloadingController(), null, Defs.CONFIG_RELOAD_INTERVAL,
-		        TimeUnit.SECONDS, executor);
+				TimeUnit.SECONDS, executor);
 		trigger.start();
 
 		return builder;
@@ -122,189 +105,106 @@ public enum GlobalConfig {
 	}
 
 	/**
-	 * 
-	 * @param timeZone
-	 */
-	public void setServerTimeZone(String timeZone) {
-		getConfig().setProperty(Options.TIMEZONE_SERVER.getName(), timeZone);
-	}
-
-	/**
-	 * 
-	 * @return
-	 */
-	public String getServerTimeZone() {
-		return getConfig().getString(Options.TIMEZONE_SERVER.getName());
-	}
-
-	/**
-	 * 
-	 * @return
-	 */
-	public boolean isMaintenanceEnabled() {
-		return getConfig().getBoolean(Options.MAINTENANCE_ENABLED.getName());
-	}
-
-	/**
-	 * 
-	 * @param enabled
-	 */
-	public void setMaintenanceEnabled(boolean enabled) {
-		getConfig().setProperty(Options.MAINTENANCE_ENABLED.getName(), enabled);
-	}
-
-	/**
-	 * 
-	 * @return Telegram bot token
-	 */
-	public String getBotToken() {
-		return getConfig().getString(Options.TELEGRAM_BOT_TOKEN.getName());
-	}
-
-	/**
-	 * 
-	 * @param token
-	 */
-	public void setBotToken(String token) {
-		getConfig().setProperty(Options.TELEGRAM_BOT_TOKEN.getName(), token);
-	}
-
-	/**
-	 * 
-	 * @return Telegram channel
-	 */
-	public String getBotChannel() {
-		return getConfig().getString(Options.TELEGRAM_CHANNEL.getName());
-	}
-
-	/**
-	 * 
-	 * @param channel
-	 */
-	public void setBotChannel(String channel) {
-		getConfig().setProperty(Options.TELEGRAM_CHANNEL.getName(), channel);
-	}
-
-	/**
-	 * 
-	 * @return SparkPost API KEY
-	 */
-	public String getSparkPostAPIKey() {
-		return getConfig().getString(Options.SPARKPOST_API_KEY.getName());
-	}
-
-	/**
-	 * 
-	 * @param API
-	 *            KEY
-	 */
-	public void setSparkPostAPIKey(String apiKey) {
-		getConfig().setProperty(Options.SPARKPOST_API_KEY.getName(), apiKey);
-	}
-
-	/**
-	 * 
-	 * @return SparkPost Emails
-	 */
-	public String getSparkPostEmails() {
-		return getConfig().getString(Options.SPARKPOST_EMAILS.getName());
-	}
-
-	/**
-	 * 
-	 * @param Emails
-	 */
-	public void setSparkPostEmails(String emails) {
-		getConfig().setProperty(Options.SPARKPOST_EMAILS.getName(), emails);
-	}
-
-	/**
-	 * 
-	 * @return SparkPost Subject
-	 */
-	public String getSparkPostSubject() {
-		return getConfig().getString(Options.SPARKPOST_SUBJECT.getName());
-	}
-
-	/**
-	 * 
-	 * @param Subject
-	 */
-	public void setSparkPostSubject(String subject) {
-		getConfig().setProperty(Options.SPARKPOST_SUBJECT.getName(), subject);
-	}
-
-	/**
-	 * 
-	 * @return SparkPost From - Domain which will be used by SparkPost to send
-	 *         notifications for errors
-	 */
-	public String getSparkPostFrom() {
-		return getConfig().getString(Options.SPARKPOST_FROM.getName());
-	}
-
-	/**
-	 * 
-	 * @param From
-	 */
-	public void setSparkPostFrom(String from) {
-		getConfig().setProperty(Options.SPARKPOST_FROM.getName(), from);
-	}
-
-	/**
-	 * 
 	 * @return Server name
 	 */
 	public String getServerName() {
 		return getConfig().getString(Options.SERVER_NAME.getName());
 	}
 
-	/**
-	 * 
-	 * @param Server
-	 *            Name
-	 */
 	public void setServerName(String serverName) {
 		getConfig().setProperty(Options.SERVER_NAME.getName(), serverName);
 	}
 
+	public void setServerTimeZone(String timeZone) {
+		getConfig().setProperty(Options.TIMEZONE_SERVER.getName(), timeZone);
+	}
+
+	public String getServerTimeZone() {
+		return getConfig().getString(Options.TIMEZONE_SERVER.getName());
+	}
+
 	/**
-	 * 
+	 * @return Telegram bot token
+	 */
+	public String getBotToken() {
+		return getConfig().getString(Options.TELEGRAM_BOT_TOKEN.getName());
+	}
+
+	public void setBotToken(String token) {
+		getConfig().setProperty(Options.TELEGRAM_BOT_TOKEN.getName(), token);
+	}
+
+	/**
+	 * @return Telegram channel
+	 */
+	public String getBotChannel() {
+		return getConfig().getString(Options.TELEGRAM_CHANNEL.getName());
+	}
+
+	public void setBotChannel(String channel) {
+		getConfig().setProperty(Options.TELEGRAM_CHANNEL.getName(), channel);
+	}
+
+	/**
+	 * @return SparkPost API KEY
+	 */
+	public String getSparkPostAPIKey() {
+		return getConfig().getString(Options.SPARKPOST_API_KEY.getName());
+	}
+
+	public void setSparkPostAPIKey(String apiKey) {
+		getConfig().setProperty(Options.SPARKPOST_API_KEY.getName(), apiKey);
+	}
+
+	/**
+	 * @return SparkPost Emails
+	 */
+	public String getSparkPostEmails() {
+		return getConfig().getString(Options.SPARKPOST_EMAILS.getName());
+	}
+
+	public void setSparkPostEmails(String emails) {
+		getConfig().setProperty(Options.SPARKPOST_EMAILS.getName(), emails);
+	}
+
+	/**
+	 * @return SparkPost Subject
+	 */
+	public String getSparkPostSubject() {
+		return getConfig().getString(Options.SPARKPOST_SUBJECT.getName());
+	}
+
+	public void setSparkPostSubject(String subject) {
+		getConfig().setProperty(Options.SPARKPOST_SUBJECT.getName(), subject);
+	}
+
+	/**
+	 * @return SparkPost From - Domain which will be used by SparkPost to send
+	 * notifications for errors
+	 */
+	public String getSparkPostFrom() {
+		return getConfig().getString(Options.SPARKPOST_FROM.getName());
+	}
+
+	public void setSparkPostFrom(String from) {
+		getConfig().setProperty(Options.SPARKPOST_FROM.getName(), from);
+	}
+
+	/**
 	 * @return Report type. For instance: SparkPost or Telegram
 	 */
 	public String getReportType() {
 		return getConfig().getString(Options.REPORT_TYPE.getName());
 	}
 
-	/**
-	 * 
-	 * @param Report
-	 *            Type
-	 */
 	public void setReportType(String reportType) {
 		getConfig().setProperty(Options.REPORT_TYPE.getName(), reportType);
-	}
-
-	public boolean isLogSqlEnabled() {
-		return getConfig().getBoolean(Options.ENABLE_LOG_SQL.getName());
-	}
-
-	/**
-	 * 
-	 * @param enabled
-	 */
-	public void setLogSqlEnabled(boolean enabled) {
-		getConfig().setProperty(Options.ENABLE_LOG_SQL.getName(), enabled);
 	}
 
 	public boolean isLogDebugEnabled() {
 		return getConfig().getBoolean(Options.ENABLE_LOG_DEBUG.getName());
 	}
 
-	/**
-	 * 
-	 * @param enabled
-	 */
 	public void setLogDebugEnabled(boolean enabled) {
 		getConfig().setProperty(Options.ENABLE_LOG_DEBUG.getName(), enabled);
 	}
