@@ -1,18 +1,18 @@
 /*
  * The MIT License
- * 
+ *
  * Copyright (c) 2015 Petar Petrov
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -21,17 +21,19 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package net.vexelon.currencybg.srv.db;
 
-import java.io.Closeable;
-import java.sql.Connection;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
+package net.vexelon.currencybg.srv.db;
 
 import net.vexelon.currencybg.srv.db.models.CurrencyData;
 import net.vexelon.currencybg.srv.db.models.CurrencySource;
 import net.vexelon.currencybg.srv.db.models.ReportData;
+
+import javax.annotation.Nonnull;
+import java.io.Closeable;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
 
 /**
  * Encapsulates the available read-write operations to and from an underlying
@@ -39,19 +41,22 @@ import net.vexelon.currencybg.srv.db.models.ReportData;
  */
 public interface DataSource extends Closeable {
 
+	@Nonnull
+	static DataSource newDataSource() {
+		return new FirestoreDataSource();
+	}
+
 	/**
 	 * Establishes connection to data source.
-	 * 
-	 * @param context
-	 * @throws DataSourceException
-	 *             If an SQL error is thrown.
+	 *
+	 * @throws DataSourceException If an SQL error is thrown.
 	 */
-	Connection connect() throws DataSourceException;
+	void connect() throws DataSourceException;
 
 	/**
 	 * Fetches a list of currencies for the current date which are after
 	 * DateTime
-	 * 
+	 *
 	 * @param timeFrom
 	 * @return
 	 * @throws DataSourceException
@@ -59,19 +64,15 @@ public interface DataSource extends Closeable {
 	String getAllCurrentRatesAfter(Date timeFrom) throws DataSourceException;
 
 	/**
-	 * Fetches a list of currencies for the current date which are after
-	 * DateTime by sourceId
-	 * 
-	 * @param sourceId
+	 * Fetches a list of currencies for the current date which are after DateTime for a {@code source}
+	 *
 	 * @param timeFrom
-	 * @return
-	 * @throws DataSourceException
 	 */
-	String getAllCurrentRatesAfter(Integer sourceId, Date timeFrom) throws DataSourceException;
+	String getAllCurrentRatesAfter(int sourceId, Date timeFrom) throws DataSourceException;
 
 	/**
 	 * Fetches a list of currencies by date
-	 * 
+	 *
 	 * @param dateFrom
 	 * @return
 	 * @throws DataSourceException
@@ -80,37 +81,25 @@ public interface DataSource extends Closeable {
 
 	/**
 	 * Fetches a list of currencies by sourceId and date
-	 * 
+	 *
 	 * @param sourceId
 	 * @param dateFrom
 	 * @return
 	 * @throws DataSourceException
 	 */
-	String getAllRates(Integer sourceId, Date dateFrom) throws DataSourceException;
+	String getAllRates(int sourceId, Date dateFrom) throws DataSourceException;
 
 	/**
-	 * Fetches a list of sources in DB by id
-	 * 
-	 * @param id
-	 * @return
-	 * @throws DataSourceException
+	 * Fetches a list of sources
+	 *
+	 * @param isActiveOnly If {@code true}, only actives sources will be fetched.
 	 */
-	CurrencySource getSourceById(int id) throws DataSourceException;
-
-	/**
-	 * Fetches a list of sources in DB
-	 * 
-	 * @return
-	 * @throws DataSourceException
-	 */
-
-	List<CurrencySource> getAllSources(boolean isActiveOnly) throws DataSourceException;
+	@Nonnull
+	Collection<CurrencySource> getAllSources(boolean isActiveOnly) throws DataSourceException;
 
 	/**
 	 * Check whether the authentication is valid
-	 * 
-	 * @param headerName
-	 * @param headerValue
+	 *
 	 * @return
 	 * @throws DataSourceException
 	 */
@@ -118,26 +107,24 @@ public interface DataSource extends Closeable {
 
 	/**
 	 * Adds Map of exchange rates in DB.
-	 * 
-	 * @param rates
-	 *            A {@link Map} of language and {@link CurrencyData} list
-	 *            values.
+	 *
+	 * @param rates A {@link Map} of language and {@link CurrencyData} list
+	 *              values.
 	 * @throws DataSourceException
 	 */
 	void addRates(Map<Integer, List<CurrencyData>> rates) throws DataSourceException;
 
 	/**
-	 * Add rates from by one source
-	 * 
-	 * @param sourceId
-	 * @param rates
+	 * Adds rates from a source
+	 *
+	 * @param rates A list of {@link CurrencyData} fetched rates.
 	 * @throws DataSourceException
 	 */
-	void addRates(List<CurrencyData> rates) throws DataSourceException;
+	void addRates(Collection<CurrencyData> rates) throws DataSourceException;
 
 	/**
 	 * Update some or all fields in cbg_sources table
-	 * 
+	 *
 	 * @param sourceId
 	 * @param source
 	 * @throws DataSourceException
@@ -145,28 +132,21 @@ public interface DataSource extends Closeable {
 	void updateSource(int sourceId, CurrencySource source) throws DataSourceException;
 
 	/**
-	 * Write an error message in DB
-	 * 
-	 * @param source
-	 *            throws an error
-	 * @param message
-	 * @throws DataSourceException
+	 * Writes an error message report in the database
+	 *
+	 * @param message Formatted message.
 	 */
-	void addReport(String message) throws DataSourceException;
+	void addReportMessage(String message) throws DataSourceException;
 
 	/**
-	 * Return information for all reports in DB
-	 * 
-	 * @return
-	 * @throws DataSourceException
+	 * @return A list of all reports waiting to be sent.
 	 */
-	List<ReportData> getReports() throws DataSourceException;
+	@Nonnull
+	Collection<ReportData> getReports() throws DataSourceException;
 
 	/**
-	 * Clear all information for the reports in DB
-	 * 
-	 * @throws DataSourceException
+	 * Remove all reports specified
 	 */
-	void deleteReports(List<ReportData> reporters) throws DataSourceException;
+	void deleteReports(@Nonnull Collection<ReportData> reports) throws DataSourceException;
 
 }
