@@ -4,20 +4,21 @@ import java.util.Date;
 
 public class ReportData {
 
-	private int id;
-	private Date createdOn;
-	private int source;
+	private int    id; // SQL
+	private String documentId; // Firestore
+	private Date   createdOn;
+	//	private int    source;
 	private String message;
 
 	public ReportData() {
 
 	}
 
-	public ReportData(int id, Date createdOn, int source, String message) {
+	public ReportData(int id, Date createdOn, String message) {
 		super();
 		this.id = id;
 		this.createdOn = createdOn;
-		this.source = source;
+		//		this.source = source;
 		this.message = message;
 	}
 
@@ -29,6 +30,14 @@ public class ReportData {
 		this.id = id;
 	}
 
+	public String getDocumentId() {
+		return documentId;
+	}
+
+	public void setDocumentId(String documentId) {
+		this.documentId = documentId;
+	}
+
 	public Date getCreatedOn() {
 		return createdOn;
 	}
@@ -37,13 +46,13 @@ public class ReportData {
 		this.createdOn = createdOn;
 	}
 
-	public int getSource() {
-		return source;
-	}
-
-	public void setSource(int source) {
-		this.source = source;
-	}
+	//	public int getSource() {
+	//		return source;
+	//	}
+	//
+	//	public void setSource(int source) {
+	//		this.source = source;
+	//	}
 
 	public String getMessage() {
 		return message;
@@ -55,8 +64,7 @@ public class ReportData {
 
 	@Override
 	public String toString() {
-		return "ReportData [id=" + id + ", createdOn=" + createdOn + ", source=" + source + ", message=" + message
-		        + "]";
+		return "ReportData{" + "id=" + id + ", documentId='" + documentId + '\'' + ", createdOn=" + createdOn
+				+ ", message='" + message + '\'' + '}';
 	}
-
 }
